Adversarial audit: type-system spirit fixes, red-team hardening, findings register
Add the authoritative partiality gate (tools/audit-type-system.sh) and record the
full findings in spec/AUDIT.md. Fix the genuine type-system "spirit" violations:
DAG Message height/seq -> BlockHeight/SeqNum, invariant-preserving Add<NonNegI64>,
radix-tree Node -> [Item; 256], reject negative deploy cost, chunker checked_sub.
Harden the network surface: wire grpc-max-recv-message-size, enforce the stream
size cap while draining, semaphore-bound dispatch, and use the dead send timeout.
Fix the spatial-matcher error swallow and reconcile six stale doc statements.

+++ casper/src/api/block_api_impl.rs
@@ -469,11 +469,11 @@ impl BlockApi for BlockApiImpl {
             .dag_message_state
             .msg_map
             .values()
-            .filter(|m| m.height >= lowest_height)
+            .filter(|m| i64::from(m.height) >= lowest_height)
             .map(|m| ValidatorBlock {
                 id: to_hash_str(m.id.as_bytes()),
                 sender: to_hash_str(m.sender.as_bytes()),
-                height: m.height,
+                height: i64::from(m.height),
                 justifications: m.parents.iter().map(|h| to_hash_str(h.as_bytes())).collect(),
                 fringe: m.fringe.iter().map(|h| to_hash_str(h.as_bytes())).collect(),
             })
